렌더링 라이브러리 전환
Win32API 렌더링 기능 제거, 엔진 리펙토링, 헤더 경로 수정

+++ Engine/private/Helpers/huruInput.cpp
@@ -1,12 +1,12 @@
 #include "Helpers/huruInput.h"
 #include "HighLevelInterface/huruApplication.h"
 
+
 extern Application application;
 
 namespace huru
 {
-
-	vector<Input::Key> Input::Keys = { };
+	vector<Input::Key> Input::Keys		= { };
 	Vector2 huru::Input::mMousePosition = Vector2::One;
 
 	int ASCII[(int)eKeyCode::End] =
@@ -96,11 +96,11 @@ namespace huru
 		mMousePosition.x = -1.0f;
 		mMousePosition.y = -1.0f;
 
-		if (mousePos.x > 0 && mousePos.x < width)
-			mMousePosition.x = mousePos.x;
+		if ((UINT)mousePos.x > 0 && (UINT)mousePos.x < width)
+			mMousePosition.x = (float)mousePos.x;
 
-		if (mousePos.y > 0 && mousePos.y < height)
-			mMousePosition.y = mousePos.y;
+		if ((UINT)mousePos.y > 0 && (UINT)mousePos.y < height)
+			mMousePosition.y = (float)mousePos.y;
 	}
 
 	void Input::clearKey()
